move link related code to service
