+++ links/storage.go
@@ -1,4 +1,4 @@
-package main
+package links
 
 import (
 	"context"
