+++ links/utils.go
@@ -1,4 +1,4 @@
-package main
+package links
 
 import (
 	"math/rand"
